% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spectra.R
\name{plot_spectra}
\alias{plot_spectra}
\title{Plot Mass Spectrometry Spectra to PDF}
\usage{
plot_spectra(full_mgf, output_folder)
}
\arguments{
\item{full_mgf}{\code{data.frame}. The data imported using the \code{\link{import_spectra}}
function.}

\item{output_folder}{A character string specifying the directory path where
the output PDF \code{"all_spectra_plot.pdf"} should be saved.}
}
\value{
No return value. The function is called for its side effect of writing
a PDF file to the specified \code{output_folder}.
}
\description{
This function iterates through unique spectrum IDs inside the mass spec data,
and exports a single multi-page PDF document containing line plots for each spectrum.
The number above each PTCR slice are the determined charge states for that slice.
}
